Combined Live Feed + Calibration Code...
This is a combination of the code Vidur was working on, which connected OpenCV directly to the Pixy's Live feed, and the OpenCV Camera Calibration code we were working on. It should show that it is recognizing the dots on the live feed, and calibrate based on each individual frames. Once you press "Enter," then it should show the undistorted images or output the calibration numbers (I'm not sure which).